requirements.txt
Django~=4.2.16
djangorestframework~=3.15.2
asgiref~=3.7.2
channels~=4.2.0
requests~=2.26.0
websocket-client~=1.6.4